math cleanup 🧮
